1，修改，修复LOGO选择控件的抖动问题。 2，修改，实现RDP服务器配置修改。 ————————————————————————
TODO：修改组名后，UI上显示的组名不会跟着改变。

+++ PRM.Core/Model/Global.cs
@@ -109,7 +109,7 @@ namespace PRM.Core.Model
 
         #region Server Data
 
-        public readonly Dictionary<long, ServerAbstract> ServerDict = new Dictionary<long, ServerAbstract>();
+        public readonly Dictionary<uint, ServerAbstract> ServerDict = new Dictionary<uint, ServerAbstract>();
 
         private void ReadServerDataFromDb()
         {
